Blueprints: Support --options, use ${{}} syntax
A vars section in the YAML manifest now specifies any user input
that's required to process the blueprint. Variables can be passed
at the command line, otherwise will prompt() for user input when
applying the blueprint.

The previous {{}} interpolation syntax has been changed to ${{}}
which seems to be what GitHub workflow YAMLs are using, as well
as Azure pipelines. It's also valid YAML (treated as a string).