Postfix: Allow postfix user to read the main.cf file
It seems as if in some cases the sendmail agent, which runs as the
postfix user, attempts to read the main.cf configuration file, and
fatals out if unsuccessful.

This PR adds the postfix user to the mail group on install, to
make sure it has the necessary permissions to do so.